moves formatting logic into its own module #379
