+++ lib/merger.ex
@@ -1,5 +1,5 @@
 defmodule Karma.Merger do
-  alias Karma.{Repo, ViewHelpers}
+  alias Karma.{Repo, Formatter}
 
   def merge(offer, document) do
     # download document
@@ -12,8 +12,8 @@ defmodule Karma.Merger do
         # get formatted data
         json =
           get_data_for_merge(offer)
-          |> format_data_for_merge()
-          |> format()
+          |> Formatter.format_data()
+          |> Formatter.format()
           |> Poison.encode!()
 
         # do merge
@@ -57,139 +57,12 @@ defmodule Karma.Merger do
     end
   end
 
-  # formats nested map of all data, prefixes and flattens
-  def format(data) do
-
-    Enum.reduce(Map.keys(data), %{}, fn(key, acc) ->
-      # prefix is "offer", or "startpack"
-      prefix = Atom.to_string(key)
-
-      Map.get(data, key) # offer, startpack, user or project
-      |> prefix_keys(prefix) # user_first_name
-      |> Map.merge(acc)
-    end)
-  end
-
-  # helper used by format function
-  defp prefix_keys(map, prefix) do
-    Enum.reduce(Map.keys(map), %{}, fn(key, acc) ->
-      prefixed_key = prefix <> "_" <> Atom.to_string(key)
-      val = Map.get(map, key)
-      # replace nulls with empty string as nulls seem to break merge
-      val = if val == nil, do: "", else: val
-      Map.put(acc, prefixed_key, val)
-    end)
-  end
-
   def get_data_for_merge(offer) do
     %{user: Map.take(Map.from_struct(Repo.get(Karma.User, offer.user_id)), user()),
       project: Map.take(Map.from_struct(Repo.get(Karma.Project, offer.project_id)), project()),
       offer: Map.take(Map.from_struct(Repo.get(Karma.Offer, offer.id)), offer()),
       startpack: Map.take(Map.from_struct(Repo.get_by(Karma.Startpack, user_id: offer.user_id)), startpack())
     }
-  end
-
-  def format_data_for_merge(data) do
-    offer_data = data.offer
-    project_data = data.project
-    startpack_data = data.startpack
-    user_data = data.user
-
-    currency_symbol =
-      case offer_data.currency do
-        "gbp" -> "£"
-        "eur" -> "€"
-        "usd" -> "$"
-      end
-
-    # format offer data
-    offer_data = %{
-      offer_data |
-      "vehicle_allowance_per_week": "#{currency_symbol}#{format_number(offer_data.vehicle_allowance_per_week)}",
-      "fee_per_day_inc_holiday": "#{currency_symbol}#{format_number(offer_data.fee_per_day_inc_holiday)}",
-      "fee_per_day_exc_holiday": "#{currency_symbol}#{format_number(offer_data.fee_per_day_exc_holiday)}",
-      "fee_per_week_inc_holiday": "#{currency_symbol}#{format_number(offer_data.fee_per_week_inc_holiday)}",
-      "fee_per_week_exc_holiday": "#{currency_symbol}#{format_number(offer_data.fee_per_week_exc_holiday)}",
-      "holiday_pay_per_day": "#{currency_symbol}#{format_number(offer_data.holiday_pay_per_week)}",
-      "holiday_pay_per_week": "#{currency_symbol}#{format_number(offer_data.holiday_pay_per_week)}",
-      "sixth_day_fee_inc_holiday": "#{currency_symbol}#{format_number(round(offer_data.sixth_day_fee_inc_holiday))}",
-      "sixth_day_fee_exc_holiday": "#{currency_symbol}#{format_number(round(offer_data.sixth_day_fee_exc_holiday))}",
-      "seventh_day_fee_inc_holiday": "#{currency_symbol}#{format_number(round(offer_data.seventh_day_fee_inc_holiday))}",
-      "seventh_day_fee_exc_holiday": "#{currency_symbol}#{format_number(round(offer_data.seventh_day_fee_exc_holiday))}",
-      "start_date": ViewHelpers.format_date(offer_data.start_date)
-    }
-    offer_data =
-      case offer_data.box_rental_required? do
-        true ->
-          %{
-            offer_data |
-            "box_rental_fee_per_week": "#{currency_symbol}#{format_number(offer_data.box_rental_fee_per_week)}",
-            "box_rental_cap": "#{currency_symbol}#{format_number(offer_data.box_rental_cap)}",
-          }
-        _not_true ->
-          offer_data
-      end
-    offer_data =
-      case offer_data.equipment_rental_required? do
-        true ->
-          %{
-            offer_data |
-            "equipment_rental_fee_per_week": "#{currency_symbol}#{format_number(offer_data.equipment_rental_fee_per_week)}",
-            "equipment_rental_cap": "#{currency_symbol}#{format_number(offer_data.equipment_rental_cap)}",
-          }
-        _not_true ->
-          offer_data
-      end
-
-    # format project data
-    project_data = %{
-      project_data |
-      "holiday_rate": ViewHelpers.format_holiday_rate(project_data.holiday_rate),
-      "start_date": ViewHelpers.format_date(project_data.start_date)
-    }
-
-    # format startpack data
-    startpack_data = %{
-      startpack_data |
-      "date_of_birth": ViewHelpers.format_date(startpack_data.date_of_birth)
-    }
-    startpack_data =
-      case offer_data.box_rental_required? do
-        true ->
-          %{
-            startpack_data |
-            "box_rental_value": "#{currency_symbol}#{format_number(startpack_data.box_rental_value)}",
-            }
-        _not_true ->
-          startpack_data
-      end
-    startpack_data =
-      case offer_data.equipment_rental_required? do
-        true ->
-          %{
-            startpack_data |
-            "equipment_rental_value": "#{currency_symbol}#{format_number(startpack_data.equipment_rental_value)}",
-            }
-        _not_true ->
-          startpack_data
-      end
-
-
-    %{
-      offer: offer_data,
-      project: project_data,
-      startpack: startpack_data,
-      user: user_data
-    }
-  end
-
-  def format_number(number) do
-    number
-    |> Integer.to_char_list
-    |> Enum.reverse
-    |> Enum.chunk(3, 3, [])
-    |> Enum.join(",")
-    |> String.reverse
   end
 
   def project do
